Merge: 빌드 에러 해결

See merge request s09-fintech-finance-sub2/S09P22B207!183

+++ Back-End/src/main/java/com/conseller/conseller/notification/NotificationService.java
@@ -16,7 +16,7 @@ public interface NotificationService {
 
     public void sendBarterRequestNotification(Long barterIdx, String title, Integer type);
 
-    public void sendGifticonNotification(Long userIdx, Integer remainDay, String gifticionName, Integer gifticonCount, Integer type);
+    public void sendGifticonNotification(Long userIdx, Integer remainDay, String gifticionName, Long gifticonCount, Integer type);
 
     public void sendNotification(Long userIdx, String title, String body);
